Backend/backend/requirements.txt
fastapi
uvicorn[standard]
redis>=5.3
asyncpg
databases[postgresql]
paho-mqtt
python-dotenv
requests